requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==24.1.0
beautifulsoup4==4.12.3
bidict==0.23.1
certifi==2019.11.28
chainlit==0.7.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
dbus-python==1.2.16
Deprecated==1.2.14
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
distro-info==0.23+ubuntu1.1
exceptiongroup==1.2.2
fastapi==0.99.1
fastapi-socketio==0.0.10
filelock==3.15.4
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.6.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==2.8
importlib_metadata==8.0.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-core==0.2.28
langchain-text-splitters==0.2.2
langsmith==0.1.96
Lazify==0.4.0
llama-cloud==0.0.11
llama-index==0.10.59
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.59
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.27
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.32
llama-index-readers-llama-parse==0.1.6
llama-index-readers-wikipedia==0.1.4
llama-parse==0.4.9
llama_cpp_python==0.2.7
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.9.1
numpy==1.26.4
openai==1.38.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
orjson==3.10.6
packaging==23.2
pandas==2.2.2
pillow==10.4.0
prisma==0.10.0
protobuf==4.25.4
pydantic==1.10.17
PyGObject==3.36.0
PyJWT==2.9.0
pypdf==4.3.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
requests-unixsocket==0.2.0
safetensors==0.4.3
simple-websocket==1.0.0
six==1.14.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.27.0
striprtf==0.0.26
syncer==2.0.3
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.13.0
tqdm==4.66.5
transformers==4.33.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unattended-upgrades==0.1
uptrace==1.26.0
urllib3==1.25.8
uvicorn==0.23.2
watchfiles==0.20.0
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
